requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
httpx>=0.27.0
gspread>=6.0.0
google-auth>=2.28.0
resend>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.7.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.9
websockets>=14.0
